deployment/process-thumb/requirements.txt
Flask==2.3.2
flask-cors==3.0.10
Flask-RESTful==0.3.9
kafka-python-ng==2.2.3
boto3==1.26.24
Pillow==9.3.0
requests==2.28.1
pydantic==1.9.0
six==1.16.0  
PyYAML
